add & use generalize assign_data_to_bins func
